+++ libcloud/compute/drivers/gce.py
@@ -1749,7 +1749,8 @@ class GCENodeDriver(NodeDriver):
                          'MULTI_IP_SUBNET']
 
     def __init__(self, user_id, key=None, datacenter=None, project=None,
-                 auth_type=None, scopes=None, credential_file=None, **kwargs):
+                 auth_type=None, scopes=None, credential_file=None,
+                 timeout=None, **kwargs):
         """
         :param  user_id: The email address (for service accounts) or Client ID
                          (for installed apps) to be used for authentication.
@@ -1783,7 +1784,14 @@ class GCENodeDriver(NodeDriver):
         :keyword  credential_file: Path to file for caching authentication
                                    information used by GCEConnection.
         :type     credential_file: ``str``
+
+        :keyword  timeout: Jobs timeout in seconds.
+        :type     timeout: ``int``
         """
+        if timeout is not None:
+            if timeout > 0:
+                self.connectionCls.timeout = timeout
+
         if not project:
             raise ValueError('Project name must be specified using '
                              '"project" keyword.')