+++ CHANGES.rst
@@ -2129,6 +2129,10 @@ Compute
   (GITHUB-371)
   [Jeroen de Korte]
 
+- Allow setting job timeout in GCENodeDriver.
+  (LIBCLOUD-720)
+  [Gabi Kazav]
+
 Storage
 ~~~~~~~